+++ packages/relax_cli/lib/src/generators/generators.dart
@@ -1,3 +1,4 @@
 export 'feature_generator.dart';
 export 'module_generator.dart';
+export 'page_generator.dart';
 export 'project_generator.dart';